pyproject.toml
[project]
name = "discordbot"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = ["discord.py", "pydantic", "pydantic-settings"]

[dependency-groups]
dev = ["ruff"]